Update getting started to remove 3rd party ref.
Per discussion in the Outreach Committee Meeting dating July 1st, it was agreed that third party content from vendors will be added to the Resources page on the website. Core documentation should not have references to vendor content. 

Once the official dockerhub for prestodb is ready, this tutorial will be moved to the Resources page